Fix compiler warnings (and one 64-bit safety bug)
Ignore-this: 6a86eeddc7b1b09b1dd7f49f5f4a6dd3

darcs-hash:20111231115953-b78f9-35e1cd49ea1e7e3148408f7d424ba163fa1e11eb

+++ src/topten.c
@@ -1,6 +1,6 @@
 /*	SCCS Id: @(#)topten.c	3.4	2000/01/21	*/
 /* Copyright (c) Stichting Mathematisch Centrum, Amsterdam, 1985. */
-/* Modified 19 Dec 2011 by Alex Smith */
+/* Modified 31 Dec 2011 by Alex Smith */
 /* NetHack may be freely redistributed.  See license for details. */
 
 #include "hack.h"
@@ -377,7 +377,7 @@ struct toptenentry *tt;
   (void)fprintf(rfile, SEP "align0=%s", aligns[1 - u.ualignbase[A_ORIGINAL]].filecode);
 
   (void)fprintf(rfile, SEP "xplevel=%d", u.ulevel); /* XP level */
-  (void)fprintf(rfile, SEP "exp=%d", u.uexp);       /* Experience points */
+  (void)fprintf(rfile, SEP "exp=%ld", (long)u.uexp); /* Experience points */
 
   (void)fprintf(rfile, SEP "mode=%s", (flags.debug ? "debug" :
 				       flags.explore ? "explore" :
